+++ frontend-admin/src/components/IntersectionObserverScroll/IntersectionObserverScroll.tsx
